+++ docs/source/conf.py
@@ -15,7 +15,7 @@ sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), "../.
 project = "Baromedr Cymru Wave 2 Dashboard"
 copyright = "Bharadwaj Raman"
 author = "Bharadwaj Raman - https://github.com/nba1990/"
-release = "0.2.1"
+release = "0.2.2"
 version = "0.2"
 
 extensions = [